@@ -1,34 +1,35 @@
-///
+//
 //  Generated code. Do not modify.
 //  source: proto/application.proto
 //
 // @dart = 2.12
-// ignore_for_file: annotate_overrides,camel_case_types,constant_identifier_names,deprecated_member_use_from_same_package,directives_ordering,library_prefixes,non_constant_identifier_names,prefer_final_fields,return_of_invalid_type,unnecessary_const,unnecessary_import,unnecessary_this,unused_import,unused_shown_name
 
-// Dart imports:
+// ignore_for_file: annotate_overrides, camel_case_types, comment_references
+// ignore_for_file: constant_identifier_names, library_prefixes
+// ignore_for_file: non_constant_identifier_names, prefer_final_fields
+// ignore_for_file: unnecessary_import, unnecessary_this, unused_import
+
 import 'dart:convert' as $convert;
 import 'dart:core' as $core;
 import 'dart:typed_data' as $typed_data;
 
 @$core.Deprecated('Use applicationDescriptor instead')
-const Application$json = const {
+const Application$json = {
   '1': 'Application',
-  '2': const [
-    const {'1': 'app_id', '3': 1, '4': 1, '5': 9, '10': 'appId'},
-    const {'1': 'app_bundle', '3': 2, '4': 1, '5': 9, '10': 'appBundle'},
-    const {'1': 'app_key', '3': 3, '4': 1, '5': 9, '10': 'appKey'},
-    const {'1': 'app_version', '3': 4, '4': 1, '5': 9, '10': 'appVersion'},
-    const {
-      '1': 'created_timestamp',
-      '3': 5,
-      '4': 1,
-      '5': 11,
-      '6': '.google.protobuf.Timestamp',
-      '10': 'createdTimestamp'
-    },
+  '2': [
+    {'1': 'app_id', '3': 1, '4': 1, '5': 9, '10': 'appId'},
+    {'1': 'app_bundle', '3': 2, '4': 1, '5': 9, '10': 'appBundle'},
+    {'1': 'app_key', '3': 3, '4': 1, '5': 9, '10': 'appKey'},
+    {'1': 'app_version', '3': 4, '4': 1, '5': 9, '10': 'appVersion'},
+    {'1': 'created_timestamp', '3': 5, '4': 1, '5': 11, '6': '.google.protobuf.Timestamp', '10': 'createdTimestamp'},
+    {'1': 'id', '3': 6, '4': 1, '5': 3, '10': 'id'},
   ],
 };
 
 /// Descriptor for `Application`. Decode as a `google.protobuf.DescriptorProto`.
 final $typed_data.Uint8List applicationDescriptor = $convert.base64Decode(
-    'CgtBcHBsaWNhdGlvbhIVCgZhcHBfaWQYASABKAlSBWFwcElkEh0KCmFwcF9idW5kbGUYAiABKAlSCWFwcEJ1bmRsZRIXCgdhcHBfa2V5GAMgASgJUgZhcHBLZXkSHwoLYXBwX3ZlcnNpb24YBCABKAlSCmFwcFZlcnNpb24SRwoRY3JlYXRlZF90aW1lc3RhbXAYBSABKAsyGi5nb29nbGUucHJvdG9idWYuVGltZXN0YW1wUhBjcmVhdGVkVGltZXN0YW1w');
+    'CgtBcHBsaWNhdGlvbhIVCgZhcHBfaWQYASABKAlSBWFwcElkEh0KCmFwcF9idW5kbGUYAiABKA'
+    'lSCWFwcEJ1bmRsZRIXCgdhcHBfa2V5GAMgASgJUgZhcHBLZXkSHwoLYXBwX3ZlcnNpb24YBCAB'
+    'KAlSCmFwcFZlcnNpb24SRwoRY3JlYXRlZF90aW1lc3RhbXAYBSABKAsyGi5nb29nbGUucHJvdG'
+    '9idWYuVGltZXN0YW1wUhBjcmVhdGVkVGltZXN0YW1wEg4KAmlkGAYgASgDUgJpZA==');
+
