@@ -15,14 +15,14 @@ import 'package:package_info_plus/package_info_plus.dart';
 import 'package:watchtower_sdk/models/app_data.dart';
 import 'package:watchtower_sdk/models/user_app_data.dart';
 import 'package:watchtower_sdk/sdk_functions.dart';
-import 'package:watchtower_sdk/session_recorder.dart';
+// import 'package:watchtower_sdk/session_recorder.dart';
 import 'package:watchtower_sdk/store.dart';
 import 'package:watchtower_sdk/watchtower_connector.dart';
 import 'package:watchtower_sdk/watchtower_logger.dart';
 import 'package:watchtower_sdk/watchtower_proto/proto/event.pb.dart';
 import 'package:watchtower_sdk/watchtower_proto/proto/events_payload.pb.dart';
-import 'package:watchtower_sdk/watchtower_proto/proto/server.pbgrpc.dart';
-import 'package:watchtower_sdk/watchtower_proto/proto/session_frame.pb.dart';
+import 'package:watchtower_sdk/watchtower_proto/proto/watchtower_api.pbgrpc.dart';
+// import 'package:watchtower_sdk/watchtower_proto/proto/session_frame.pb.dart';
 
 var logger = getLogger("watchtower_sdk");
 
@@ -43,24 +43,24 @@ class Watchtower {
   static String? idfa = "";
   static String sessionId = "";
   static late WatchtowerConnector watchtowerConnector;
-  static final sessionRecoreder = SessionRecorder();
+  // static final sessionRecoreder = SessionRecorder();
   static late bool isSessionRecorderEnabeled;
 
   // Init SDK singletone instance
   static Future<void> init({
     required String appId,
     required String appKey,
-    GlobalKey? repaintBoundary,
+    // GlobalKey? repaintBoundary,
     String host = 'watchtower.cyou',
-    int port = 8008,
-    bool enableSessionRecorder = false,
+    int port = 50053,
+    // bool enableSessionRecorder = false,
     bool useTls = true,
-    int sessionRecordIntervalInMs = 100,
+    // int sessionRecordIntervalInMs = 100,
   }) async {
-    validateWatchtowerParams(
-        enableSessionRecorder: enableSessionRecorder,
-        repaintBoundary: repaintBoundary,
-        sessionRecordIntervalInMs: sessionRecordIntervalInMs);
+    // validateWatchtowerParams(
+    //     enableSessionRecorder: enableSessionRecorder,
+    //     repaintBoundary: repaintBoundary,
+    //     sessionRecordIntervalInMs: sessionRecordIntervalInMs);
     await getDeviceInfo();
 
     // If SDK already initialized skip init process
@@ -104,59 +104,59 @@ class Watchtower {
     sendAppStartEvent();
     await _processLocaleStoreSavedEvents();
 
-    isSessionRecorderEnabeled = enableSessionRecorder;
-    if (repaintBoundary != null && enableSessionRecorder) {
-      _initSessionrecorder(repaintBoundary, sessionRecordIntervalInMs);
-    }
+    // isSessionRecorderEnabeled = enableSessionRecorder;
+    // if (repaintBoundary != null && enableSessionRecorder) {
+    //   _initSessionrecorder(repaintBoundary, sessionRecordIntervalInMs);
+    // }
   }
 
-  static void validateWatchtowerParams(
-      {bool? enableSessionRecorder, GlobalKey? repaintBoundary, int? sessionRecordIntervalInMs}) {
-    if (enableSessionRecorder != null && enableSessionRecorder) {
-      assert(repaintBoundary != null);
-    }
+  // static void validateWatchtowerParams(
+  //     {bool? enableSessionRecorder, GlobalKey? repaintBoundary, int? sessionRecordIntervalInMs}) {
+  //   if (enableSessionRecorder != null && enableSessionRecorder) {
+  //     assert(repaintBoundary != null);
+  //   }
 
-    if (sessionRecordIntervalInMs != null) {
-      assert(sessionRecordIntervalInMs >= 100, "Session record interval should be greater or equal to 100 ms");
-      assert(sessionRecordIntervalInMs <= 10000, "Session record interval should be less or equal to 10000 ms");
-      assert(repaintBoundary != null);
-    }
-  }
+  //   if (sessionRecordIntervalInMs != null) {
+  //     assert(sessionRecordIntervalInMs >= 100, "Session record interval should be greater or equal to 100 ms");
+  //     assert(sessionRecordIntervalInMs <= 10000, "Session record interval should be less or equal to 10000 ms");
+  //     assert(repaintBoundary != null);
+  //   }
+  // }
 
-  static Future<void> _initSessionrecorder(GlobalKey repaintBoundary, int sessionRecordIntervalInMs) async {
-    logger.d("Init sessions recorder");
-    sessionRecoreder.init(
-        sessionId: sessionId, repaintBoundary: repaintBoundary, pixelRatio: 1.0, interval: sessionRecordIntervalInMs);
+  // static Future<void> _initSessionrecorder(GlobalKey repaintBoundary, int sessionRecordIntervalInMs) async {
+  //   logger.d("Init sessions recorder");
+  //   sessionRecoreder.init(
+  //       sessionId: sessionId, repaintBoundary: repaintBoundary, pixelRatio: 1.0, interval: sessionRecordIntervalInMs);
 
-    // Subsctibe to a screenshot local store stream to save screenshot if GRPC server unavailable
-    logger.d("Enable saving session frames to local store");
-    sessionRecoreder.screenshotLocalStoreStreamController.stream.listen((pngData) {
-      DataStore().saveSessionFrame(
-          sessionId: sessionId,
-          frame: SessionFrame(
-              appId: appData.appId,
-              appBundle: appData.appBundle,
-              appKey: appData.appKey,
-              userId: userAppData.userId,
-              sessionId: sessionId,
-              frameTimestamp: currentTimeStamp(),
-              frame: pngData));
-    });
+  //   // Subsctibe to a screenshot local store stream to save screenshot if GRPC server unavailable
+  //   logger.d("Enable saving session frames to local store");
+  //   sessionRecoreder.screenshotLocalStoreStreamController.stream.listen((pngData) {
+  //     DataStore().saveSessionFrame(
+  //         sessionId: sessionId,
+  //         frame: SessionFrame(
+  //             appId: appData.appId,
+  //             appBundle: appData.appBundle,
+  //             appKey: appData.appKey,
+  //             userId: userAppData.userId,
+  //             sessionId: sessionId,
+  //             frameTimestamp: currentTimeStamp(),
+  //             frame: pngData));
+  //   });
 
-    await _startSessionRecordTransmition();
+  // await _startSessionRecordTransmition();
 
-    logger.d("Check session frames saved to local store");
-    ResponseStream<SessionFrameAcceptStatus> rsps =
-        watchtowerConnector.stub.postSessionRecord(DataStore().getAllSessionsframes());
-    try {
-      await for (var item in rsps) {
-        logger.d("Stream response: $item");
-        DataStore().deleteSessionRecordById(item.frameId);
-      }
-    } on GrpcError catch (_) {
-      logger.d("Cached frames send done");
-    }
-  }
+  //   logger.d("Check session frames saved to local store");
+  //   ResponseStream<SessionFrameAcceptStatus> rsps =
+  //       watchtowerConnector.stub.postSessionRecord(DataStore().getAllSessionsframes());
+  //   try {
+  //     await for (var item in rsps) {
+  //       logger.d("Stream response: $item");
+  //       DataStore().deleteSessionRecordById(item.frameId);
+  //     }
+  //   } on GrpcError catch (_) {
+  //     logger.d("Cached frames send done");
+  //   }
+  // }
 
   static Future<void> _processLocaleStoreSavedEvents() async {
     logger.d("Check is inmemory event exists");
@@ -284,40 +284,40 @@ class Watchtower {
   }
 
   static void _onWatchtowerConnectionStateChanged(bool state) {
-    if (isSessionRecorderEnabeled) {
-      logger.d("Update session recorder send to watchtower state tp $state");
-      sessionRecoreder.isSendToWatchtowerEnabled = state;
+    //   if (isSessionRecorderEnabeled) {
+    //     logger.d("Update session recorder send to watchtower state tp $state");
+    //     sessionRecoreder.isSendToWatchtowerEnabled = state;
 
-      // If connection to watchtower restore
-      if (state == true) {
-        logger.d("reassign stream reader");
-        if (!sessionRecoreder.screenshotStreamController.hasListener) {
-          _startSessionRecordTransmition();
-        }
-      }
-    }
+    //     // If connection to watchtower restore
+    //     if (state == true) {
+    //       logger.d("reassign stream reader");
+    //       if (!sessionRecoreder.screenshotStreamController.hasListener) {
+    //         _startSessionRecordTransmition();
+    //       }
+    //     }
+    //   }
   }
 
-  static Future<void> _startSessionRecordTransmition() async {
-    logger.d("Start session record transmition");
-    try {
-      ResponseStream<SessionFrameAcceptStatus> rsps = watchtowerConnector.stub.postSessionRecord(
-          sessionRecoreder.screenshotStreamController.stream.map(((Uint8List? pngData) => SessionFrame(
-              appId: appData.appId,
-              appBundle: appData.appBundle,
-              appKey: appData.appKey,
-              userId: userAppData.userId,
-              sessionId: sessionId,
-              frameTimestamp: currentTimeStamp(),
-              frame: pngData))),
-          options: CallOptions(timeout: const Duration(hours: 1)));
-      await for (var item in rsps) {
-        logger.d("Stream response record transmission: $item");
-      }
-    } catch (e) {
-      logger.e("Stream error: $e");
-    }
-  }
+  // static Future<void> _startSessionRecordTransmition() async {
+  //   logger.d("Start session record transmition");
+  //   try {
+  //     ResponseStream<SessionFrameAcceptStatus> rsps = watchtowerConnector.stub.postSessionRecord(
+  //         sessionRecoreder.screenshotStreamController.stream.map(((Uint8List? pngData) => SessionFrame(
+  //             appId: appData.appId,
+  //             appBundle: appData.appBundle,
+  //             appKey: appData.appKey,
+  //             userId: userAppData.userId,
+  //             sessionId: sessionId,
+  //             frameTimestamp: currentTimeStamp(),
+  //             frame: pngData))),
+  //         options: CallOptions(timeout: const Duration(hours: 1)));
+  //     await for (var item in rsps) {
+  //       logger.d("Stream response record transmission: $item");
+  //     }
+  //   } catch (e) {
+  //     logger.e("Stream error: $e");
+  //   }
+  // }
 
   static Future<void> _postBatchEvents({required List<Event> events}) async {
     logger.d("Post batch events");
